requirements.txt
flask_socketio
eventlet
pyserial
